Refactor the code to allow separated buffers for each file
The code for calling each of the three tools (qmlscene, qmllint and
qmltestrunner) was refactored to use `start-process` instead of
`compile`. The motivation was to allow more control such that each
tool call with a file has its own buffer. This allows, for instance,
calling qmlscene with different QML files.

Also format the code using elisp-autofmt.